+++ src/model.js
@@ -1,12 +1,6 @@
-/*
-  model.js
-  This file is required. It must export a class with at least one public function called `getData`
-  Documentation: http://koopjs.github.io/docs/specs/provider/
-*/
 const { google } = require('googleapis')
 const analytics = google.analyticsreporting({ version: 'v4' })
 const hash = require('object-hash')
-const _ = require('lodash')
 const {
   goog: {
     analyticsTimezone,
@@ -21,11 +15,10 @@ const {
 } = require('config')
 const { CodedError } = require('./error')
 const { TIME_DIMENSIONS, providerParamToGoogle, googleParamToProvider } = require('./constants-and-lookups')
-const validateParams = require('./param-validation')
+const validateParams = require('./validation')
 const backfillTimeseriesFeatures = require('./backfill-timeseries-features')
-const { timeDimensionToTimestamp } = require('./time')
+const { timeDimensionToTimestamp } = require('./validation/time')
 const { transformDimensionPredicate, transformMetricPredicate } = require('./transform')
-const countries = require('../countries')
 const scopes = 'https://www.googleapis.com/auth/analytics.readonly'
 const MAX_RECORD_COUNT = 10000
 const googAnalyticsTimezone = process.env.GOOGLE_ANALYTICS_TIMEZONE || analyticsTimezone
@@ -42,20 +35,15 @@ function Model () {}
 Model.prototype.getData = function (req, callback) {
   if (!googViewId || !googAnalyticsTimezone || !googPrivateKey || !googClientEmail) return callback(new CodedError('Environment variables required for accessing Google Analytics are missing.', 500))
 
-  // Get the validated request params (validate and set defaults for missing optionals)
-  const params = validateParams(req)
+  const { error, value: params } = validateParams(req)
 
-  // Reject request if parameter
-  if (params.error) return callback(params.error)
+  if (error) return callback(error)
 
   // Leverage outFields to maintain desired set of output attributes
-  req.query.outFields = [].concat(params.value.outFields, ['OBJECTID']).join(',')
-
-  // If dimensioned by country, add the ISO code dimension so we can link to geometry
-  if (req.params.method === 'query' && params.value.dimension.includes('country') && !params.value.dimension.includes('countryIsoCode')) params.value.dimension.push('countryIsoCode')
+  req.query.outFields = [].concat(params.outFields, ['OBJECTID']).join(',')
 
   // Query google analytics
-  analytics.reports.batchGet(prepareRequestBody(params.value))
+  analytics.reports.batchGet(prepareRequestBody(params))
     .then(result => {
       // If a full set of results where unable to be returned, reject the request, otherwise the time series will be incomplete or the statistics will be wrong
       if (result.data.reports[0].data.rowCount && result.data.reports[0].data.rowCount > result.data.reports[0].data.rows.length) return callback(new CodedError('Request exceeds maximum page-size from Google Analytics.  Reduce time-range or alter dimension.', 400))
@@ -70,19 +58,19 @@ Model.prototype.getData = function (req, callback) {
       }
 
       geojson.metadata = {
-        name: `${params.value.metric.join(', ')} by ${params.value.dimension.join(', ')}`,
-        title: `${params.value.metric.join(', ')} by ${params.value.dimension.join(', ')}`,
+        name: `${params.metric.join(', ')} by ${params.dimension.join(', ')}`,
+        title: `${params.metric.join(', ')} by ${params.dimension.join(', ')}`,
         description: 'This provider converts Koop request parameters to parameters required by Google Analytics, and translates the response to GeoJSON for use in Koop output services',
         maxRecordCount: MAX_RECORD_COUNT
       }
 
-      if (params.value.dimension.includes('country') || params.value.dimension.includes('countryIsoCode')) geojson.metadata.geometryType = 'MultiPolygon'
+      if (req.query.where) geojson.filtersApplied = { where: true }
 
-      if (shouldBackfill(params.value.dimension)) {
+      if (shouldBackfill(params.dimension)) {
         geojson.features = backfillTimeseriesFeatures({
-          startDate: params.value.time.startDate,
-          endDate: params.value.time.endDate,
-          interval: getTimeDimension(params.value.dimension),
+          startDate: params.time.startDate,
+          endDate: params.time.endDate,
+          interval: getTimeDimension(params.dimension),
           timezone: analyticsTimezone,
           geojson
         })
@@ -106,7 +94,6 @@ Model.prototype.createKey = function (req) {
   }
 
   let key = req.url.split('/')[1]
-  if (req.params.host) key = `${key}::${req.params.host}`
   if (req.params.id) key = `${key}::${req.params.id}`
   return key
 }
@@ -173,15 +160,7 @@ function translate (metadata, data) {
   }
 }
 
-/**
- * Convert a Google Analytics Reporting row to a GeoJSON feature
- * @param {object} input - google analytics report "row"
- * @param {array} dimensionKeys - array of dimension keys/labels (e.g., "day")
- * @param {array} metricsMetadata - array of metric keys/labels (e.g., "views")
- * @returns {object} geojson feature without geometry
- */
 function formatFeature (input, metadata) {
-  let geometry = null
   const properties = {}
   const dimensions = input.dimensions || []
 
@@ -198,16 +177,10 @@ function formatFeature (input, metadata) {
     properties[metadata.metrics[i].name] = (metadata.metrics[i].type === 'INTEGER') ? Number(metric) : metric
   })
 
-  // Support geometry for country dimensioning.  If the properties include countryIsoCode, we can use it to pull geometry from a country.json file
-  if (properties['countryIsoCode']) {
-    const country = _.find(countries.features, ['properties.ISO_A2', properties.countryIsoCode]) || {}
-    geometry = country.geometry
-  }
-
   return {
     type: 'Feature',
     properties,
-    geometry
+    geometry: null
   }
 }